+++ src/kumori-main.ts
@@ -2,7 +2,7 @@
 
 import * as program from 'commander'
 import * as logger from './logger'
-import { run } from './utils'
+import { run, executeProgram } from './utils'
 
 import { workspace } from './workspace';
 
@@ -14,6 +14,7 @@ module.exports = function (argv: string[]): void {
         .version(pkg.version)
         .command('component', 'Manages components')
         .command('deployment', 'Manages deployments')
+        .command('project', 'Manages projects')
         .command('resource', 'Manages resources')
         .command('runtime', 'Manages runtimes')
         .command('service', 'Manages services')
@@ -23,16 +24,18 @@ module.exports = function (argv: string[]): void {
     program
         .command('init')
         .description('Populates the current folder with the worspace folders structure')
-        .action(() => {
+        .option('-t, --template <template>', 'The workspace template', '@kumori/workspace')
+        .action(({template}) => {
             run(async () => {
                 if (workspace.isValidWorkspace()) {
                     logger.info("Workspace already created")
                 } else {
-                    workspace.init()
+                    workspace.init(template)
                 }
             })
         })
 
-    program
-        .parse(process.argv)
+        executeProgram(program)
+        // program
+        //     .parse(process.argv)
 }