+++ src/kumori-resource.ts
@@ -1,11 +1,11 @@
 import * as program from 'commander'
 import * as logger from './logger'
 import { workspace } from './workspace/index'
-import { run } from './utils'
+import { run, executeProgram } from './utils'
 
-let defaultDomain = workspace.components.configManager.config.domain
-let defaultTemplate = workspace.components.configManager.config.resource.template
-let defaultStamp = workspace.components.configManager.config.defaultStamp.name
+let defaultDomain = workspace.configManager.config.domain
+let defaultTemplate = workspace.configManager.config.resource.template
+let defaultStamp = workspace.configManager.config.defaultStamp.name
 
 program
     .command('add <name>')
@@ -63,4 +63,5 @@ program
         })
     })
 
-program.parse(process.argv);
+executeProgram(program)
+// program.parse(process.argv);