+++ src/kumori-set.ts
@@ -1,7 +1,7 @@
 import * as program from 'commander'
 import * as logger from './logger'
 import { workspace } from './workspace';
-import { run } from './utils'
+import { run, executeProgram } from './utils'
 
 program
     .command('domain <domain>')
@@ -9,10 +9,11 @@ program
     .action((domain) => {
         run(async () => {
             logger.info(`Seting default domain to ${domain}`)
-            await workspace.config.setDefaultDomain(domain)
+            await workspace.configManager.setDefaultDomain(domain)
             logger.info(`New domain set`)
         })
     })
 
 
-program.parse(process.argv);
+executeProgram(program)
+// program.parse(process.argv);