+++ src/workspace/element-manager.ts
@@ -1,6 +1,6 @@
 import { WorkspaceConfigManager, StampConfig } from './workspace-manager'
 import { Domain, Path, Version } from './types'
-import { getJSON, checkParameter, checkIsNumber } from './utils'
+import { getJSON, checkName, checkParameter, checkIsNumber } from './utils'
 import * as path from 'path'
 import * as fs from 'fs-extra'
 import { AdmissionClient } from '@kumori/admission-client'
@@ -112,6 +112,8 @@ export class ElementManager {
         } catch(error) {
             if (error.code && (error.code.localeCompare('ECONNREFUSED') == 0)) {
                 error.message = `Connection to ${stamp} refused`
+            } else if (error.message && (error.message.indexOf("Authentication error") != -1)) {
+                error.message = `Authentication failure. Maybe your token expired or is not set. Use "kumori stamp update ${stamp} -t <TOKEN>" command to change it. Access your user settings in the platform dashboard to get a valid token.`
             } else {
                 error.message = `Unable to connect to ${stamp}`
             }
@@ -133,5 +135,9 @@ export class ElementManager {
     protected async _removeElement(name: string, domain?: Domain): Promise<void> {
         let elemPath = this._getElementFolder(name, domain)
         await fs.remove(elemPath)
+    }
+
+    protected _checkName(name: string): void {
+        return checkName(name);
     }
 }